Cleanup source code, add `/api/badge`
