--- .travis.yml
@@ -1,8 +1,0 @@
-sudo: required
-language: node_js
-node_js:
-- '8'
-
-install:
-- npm install
-
