--- api/locales
@@ -1,1 +1,0 @@
-../locales/
