+++ api/routes/index.js
@@ -2,13 +2,12 @@ const express = require('express');
 const router = express.Router();
 const request = require('request');
 
-const config = require('../config');
+const config = require('../../lib/config');
 const { badge } = require('../lib/badge');
 
 const sanitize = require('sanitize');
 
 router.get('/', function(req, res) {
-  res.setLocale(config.locale);
   res.render('index', { community: config.community,
                         tokenRequired: !!config.inviteToken,
                         recaptchaSiteKey: config.recaptchaSiteKey });
@@ -107,56 +106,6 @@ router.post('/invite', function(req, res) {
       isFailed: true
     });
   }
-});
-
-router.get('/badge.svg', (req, res) => {
-  request.get({
-    url: 'https://'+ config.slackUrl + '/api/users.list',
-    qs: {
-      token: config.slacktoken,
-      presence: true
-    }
-  }, function(err, httpResponse, body) {
-    try {
-      body = JSON.parse(body);
-    } catch(e) {
-      return res.status(404).send('');
-    }
-    if (!body.members) {
-      return res.status(404).send('');
-    }
-
-    const members = body.members.filter(function(m) {
-      return !m.is_bot;
-    });
-    const total = members.length;
-    const presence = members.filter(function(m) {
-      return m.presence === 'active';
-    }).length;
-
-    const hexColor = /^([A-Fa-f0-9]{6}|[A-Fa-f0-9]{3})$/;
-    sanitize.middleware.mixinFilters(req);
-
-    res.type('svg');
-    res.set('Cache-Control', 'max-age=0, no-cache');
-    res.set('Pragma', 'no-cache');
-    res.send(
-        badge(
-            presence,
-            total,
-            req.queryPattern('colorA', hexColor),
-            req.queryPattern('colorB', hexColor)
-        )
-    );
-  });
-});
-
-// Force it to go where I want!
-router.get('*', function(req, res) {
-  res.setLocale(config.locale);
-  res.render('index', { community: config.community,
-                        tokenRequired: !!config.inviteToken,
-                        recaptchaSiteKey: config.recaptchaSiteKey });
 });
