--- aws/api-swagger-template.json
@@ -1,41 +1,0 @@
-{
-	"swagger": "2.0",
-	"info": {
-		"version": "1.0.0",
-		"title": "%ApiTitle%"
-	},
-	"basePath": "/",
-	"schemes": ["https"],
-	"consumes": ["application/json"],
-	"produces": ["application/json"],
-	"paths": {
-		"/": {
-			"x-amazon-apigateway-any-method": {
-				"x-amazon-apigateway-integration": {
-					"uri": "arn:aws:apigateway:%Region%:lambda:path/2015-03-31/functions/arn:aws:lambda:%Region%:%AccountId%:function:${stageVariables.ExpressFunction}/invocations",
-					"passthroughBehavior": "when_no_match",
-					"httpMethod": "POST",
-					"type": "aws_proxy"
-				}
-			}
-		},
-		"/{proxy+}": {
-			"x-amazon-apigateway-any-method": {
-				"parameters": [
-					{
-						"name": "proxy",
-						"in": "path",
-						"required": true,
-						"type": "string"
-					}
-				],
-				"x-amazon-apigateway-integration": {
-					"uri": "arn:aws:apigateway:%Region%:lambda:path/2015-03-31/functions/arn:aws:lambda:%Region%:%AccountId%:function:${stageVariables.ExpressFunction}/invocations",
-					"httpMethod": "POST",
-					"type": "aws_proxy"
-				}
-			}
-		}
-	},
-	"x-amazon-apigateway-binary-media-types": ["*/*"]
-}
