--- aws/brick/.gitignore
@@ -1,2 +1,0 @@
-lambda.zip
-build/
