--- aws/brick/brick.yaml
@@ -1,10 +1,0 @@
-name: lsuss.slack_inviter
-version: v0.1.2
-license: MIT
-description: |
-  A serverless web application to invite users into your Slack
-main: :brick
-require:
-  iclab.simple: v0.2.0
-assets:
-- lambda.zip
