--- aws/config.example.sh
@@ -1,15 +1,0 @@
-#!/bin/bash
-
-StackName=SlackInviter
-
-S3BucketArtifacts=your-artifact-bucket
-S3PrefixArtifacts=cloudformation/slack-invite-automation
-
-CommunityName="Your Community Name"
-SlackUrl=yourcommunity.slack.com
-SlackToken=xoxp-xxx-xxx-xxx-xxx
-InviteToken=
-RecaptchaSiteKey=
-RecaptchaSecretKey=
-Locale=en
-
