--- azuredeploy.json
@@ -1,169 +1,0 @@
-{
-    "$schema": "https://schema.management.azure.com/schemas/2015-01-01/deploymentTemplate.json#",
-    "contentVersion": "1.0.0.0",
-    "parameters": {
-        "siteName": {
-            "type": "string"
-        },
-        "hostingPlanName": {
-            "type": "string"
-        },
-        "siteLocation": {
-            "type": "string",
-            "defaultValue": "West US"
-        },
-        "sku": {
-            "type": "string",
-            "allowedValues": [
-                "F1",
-                "D1",
-                "B1",
-                "S1"
-            ],
-            "defaultValue": "F1"
-        },
-        "workerSize": {
-            "type": "string",
-            "allowedValues": [
-                "0",
-                "1",
-                "2"
-            ],
-            "defaultValue": "0"
-        },
-        "repoURL": {
-            "type": "string",
-            "defaultValue": "https://github.com/outsideris/slack-invite-automation.git"
-        },
-        "branch": {
-            "type": "string",
-            "defaultValue": "master"
-        },
-        "communityName": {
-            "type": "string"
-        },
-        "slackUrl": {
-            "type": "string"
-        },
-        "slackToken": {
-            "type": "string"
-        },
-        "inviteToken": {
-            "type": "string"
-        },
-        "recaptchaSite": {
-            "type": "string",
-            "defaultValue": ""
-        },
-        "recaptchaSecret": {
-            "type": "string",
-            "defaultValue": ""
-        },
-        "locale": {
-            "type": "string",
-            "defaultValue": "en",
-            "allowedValues": [
-                "cs",
-                "de",
-                "en",
-                "es",
-                "fr",
-                "it",
-                "ja",
-                "ko",
-                "nl",
-                "pl",
-                "pt",
-                "pt-BR",
-                "tr",
-                "zh-CN",
-                "zh-TW"
-            ]
-        },
-        "subpath": {
-            "type": "string",
-            "defaultValue": "/"
-        }
-    },
-    "variables": {},
-    "resources": [
-        {
-            "apiVersion": "2015-08-01",
-            "name": "[parameters('hostingPlanName')]",
-            "type": "Microsoft.Web/serverfarms",
-            "location": "[parameters('siteLocation')]",
-            "properties": {
-                "name": "[parameters('hostingPlanName')]"
-            },
-            "sku": {
-                "name": "[parameters('sku')]"
-            }
-        },
-        {
-            "apiVersion": "2015-08-01",
-            "name": "[parameters('siteName')]",
-            "type": "Microsoft.Web/sites",
-            "location": "[parameters('siteLocation')]",
-            "tags": {
-                "[concat('hidden-related:', resourceGroup().id, '/providers/Microsoft.Web/serverfarms/', parameters('hostingPlanName'))]": "empty"
-            },
-            "dependsOn": [
-                "[resourceId('Microsoft.Web/serverfarms', parameters('hostingPlanName'))]"
-            ],
-            "properties": {
-                "name": "[parameters('siteName')]",
-                "serverFarmId": "[parameters('hostingPlanName')]"
-            },
-            "resources": [
-                {
-                    "apiVersion": "2015-08-01",
-                    "name": "web",
-                    "type": "config",
-                    "dependsOn": [
-                        "[resourceId('Microsoft.Web/Sites', parameters('siteName'))]"
-                    ],
-                    "properties": {}
-                },
-                {
-                    "apiVersion": "2015-08-01",
-                    "name": "appsettings",
-                    "type": "config",
-                    "dependsOn": [
-                        "[resourceId('Microsoft.Web/Sites', parameters('siteName'))]"
-                    ],
-                    "properties": {
-                        "COMMUNITY_NAME": "[parameters('communityName')]",
-                        "SLACK_URL": "[parameters('slackUrl')]",
-                        "SLACK_TOKEN": "[parameters('slackToken')]",
-                        "INVITE_TOKEN": "[parameters('inviteToken')]",
-                        "RECAPTCHA_SITE": "[parameters('recaptchaSite')]",
-                        "RECAPTCHA_SECRET": "[parameters('recaptchaSecret')]",
-                        "LOCALE": "[parameters('locale')]",
-                        "SUBPATH": "[parameters('subpath')]",
-                        "WEBSITE_NODE_DEFAULT_VERSION": "8.9.4"
-                    }
-                },
-                {
-                    "apiVersion": "2015-08-01",
-                    "name": "web",
-                    "type": "sourcecontrols",
-                    "dependsOn": [
-                        "[resourceId('Microsoft.Web/Sites', parameters('siteName'))]",
-                        "[concat('Microsoft.Web/Sites/', parameters('siteName'), '/config/web')]"
-                    ],
-                    "properties": {
-                        "RepoUrl": "[parameters('repoURL')]",
-                        "branch": "[parameters('branch')]",
-                        "IsManualIntegration": true
-                    }
-                }
-            ]
-        }
-    ],
-    "outputs": {
-        "siteUri": {
-            "type": "string",
-            "value": "[concat('https://',reference(resourceId('Microsoft.Web/sites', parameters('siteName'))).hostNames[0])]"
-        }
-    }
-}
