--- iisnode.yaml
@@ -1,3 +1,0 @@
-node_env: production
-loggingEnabled: true
-logDirectory: iisnode
