--- index.html
@@ -1,1 +1,0 @@
-<h1>hi 3</h1>
