+++ lib/config.js
@@ -16,7 +16,4 @@ module.exports = {
   // an optional security measure - if both are set, then recaptcha will be used.
   recaptchaSiteKey: process.env.RECAPTCHA_SITE || null,
   recaptchaSecretKey: process.env.RECAPTCHA_SECRET || null,
-  // default locale
-  locale: process.env.LOCALE || "en",
-  subpath: process.env.SUBPATH || "/"
 };
