--- locales/cs.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Zadej e-mail a staň se členem skupiny %s na Slacku!",
-  "ENTER_EMAIL": "Zadej svůj e-mail",
-  "ENTER_TOKEN": "Zadej přihlašovací kód z pozvánky",
-  "TITLE": "Staň se členem skupiny %s na Slacku!"
-}
