--- locales/de.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Gib deine E-Mail Adresse an, um %s auf Slack beizutreten!",
-  "ENTER_EMAIL": "Deine E-Mail Adresse",
-  "ENTER_TOKEN": "Gib deinen Zugangscode an",
-  "TITLE": "Tritt %s auf Slack bei!"
-}
