--- locales/es.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "¡Ingresa tu dirección de email debajo para unirte a %s en Slack!",
-  "ENTER_EMAIL": "Ingresa tu dirección de email",
-  "ENTER_TOKEN": "Ingresa el código de invitación",
-  "TITLE": "¡Únete a la comunidad de %s en Slack!"
-}
