--- locales/fr.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Entrez votre adresse mail pour rejoindre le groupe %s sur Slack!",
-  "ENTER_EMAIL": "Entrez votre adresse mail",
-  "ENTER_TOKEN": "Entrez votre token d'invitation",
-  "TITLE": "Rejoignez le groupe %s sur Slack!"
-}
