--- locales/it.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Inseirisci qui sotto la tua email per unirti a %s su Slack!",
-  "ENTER_EMAIL": "Inserisci il tuo indirizzo e-mail",
-  "ENTER_TOKEN": "Inserisci il token d'invito che ti è stato fornito",
-  "TITLE": "Unisciti a %s su Slack!"
-}
