--- locales/ja.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Slackで %s に参加するためにメールアドレスを入力してください！",
-  "ENTER_EMAIL": "メールアドレスを入力してください",
-  "ENTER_TOKEN": "あなたが取得した招待用トークンを入力してください。",
-  "TITLE": "Slackで %s コミュニティに参加しましょう！"
-}
