--- locales/ko.json
@@ -1,6 +1,0 @@
-{
-	"HEADER": "Slack의 %s에 가입하기 위한 이메일 주소를 입력하세요!",
-	"ENTER_EMAIL": "이메일 주소를 입력하세요",
-	"ENTER_TOKEN": "초대 토큰을 입력하세요",
-	"TITLE": "Slack의 %s 커뮤니티에 참여하세요!"
-}
