--- locales/nl.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Voer uw email adres in om deel te nemen aan %s op Slack!",
-  "ENTER_EMAIL": "Voer uw email adres in",
-  "ENTER_TOKEN": "Voer het token in dat u kreeg bij uw uitnodiging",
-  "TITLE": "Neem deel aan de %s gemeenschap op Slack!"
-}
