--- locales/pl.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Podaj swój adres email, aby dołączyć do %s na Slacku!",
-  "ENTER_EMAIL": "Podaj swój adres email",
-  "ENTER_TOKEN": "Podaj swój kod zaproszenia",
-  "TITLE": "Dołącz do społeczności %s na Slacku!"
-}
