--- locales/pt-BR.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Digite seu email abaixo para se juntar a %s no Slack!",
-  "ENTER_EMAIL": "Insira seu endereço de e-mail",
-  "ENTER_TOKEN": "Digite seu Token de convite que foi dado a você",
-  "TITLE": "Junte-se a comunidade %s no Slack!"
-}
