--- locales/pt.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Introduza o seu email para participar no Slack de %s!",
-  "ENTER_EMAIL": "Insira o seu endereço de e-mail",
-  "ENTER_TOKEN": "Digite o convite que lhe foi entregue",
-  "TITLE": "Junte-se à comunidade Slack de %s!"
-}
