--- locales/ru.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Напишите свой E-Mail чтобы присоединиться к %s в Slack!",
-  "ENTER_EMAIL": "Введите e-mail на него будет отправлено приглашение",
-  "ENTER_TOKEN": "Введите ключ-приглашение",
-  "TITLE": "Присоиденяйтесь к %s в Slack!"
-}
