--- locales/tr.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "Slack'te %s topluluğuna katılmak için e-mailinizi giriniz.",
-  "ENTER_EMAIL": "E-mail adresinizi giriniz.",
-  "ENTER_TOKEN": "Size verilen davet token'ınızı giriniz.",
-  "TITLE": "Slack'te %s topluluğuna katılın"
-}
