--- locales/zh-CN.json
@@ -1,6 +1,0 @@
-{
-  "HEADER": "以下请输入您的电子邮箱加入 %s 的 Slack！",
-  "ENTER_EMAIL": "输入您的电子邮箱地址",
-  "ENTER_TOKEN": "输入提供予您的邀请令牌",
-  "TITLE": "加入 %s 的 Slack 社群！"
-}
