--- manifest.yml
@@ -1,29 +1,0 @@
-applications:
-- name: slack-invite-automation
-  host: slack-invite-automation
-  command: bin/www
-  instances: 1
-  memory: 128M
-  disk_quota: 128M
-  env:
-    # your community or team name to display on join page
-    COMMUNITY_NAME: SocketIO
-
-    # your slack team url (ex: socketio.slack.com)
-    SLACK_URL: socketio.slack.com
-
-    # access token of slack
-    # You can generate it in https://api.slack.com/web#auth
-    # You should generate the token in admin user, not owner.
-    # If you generate the token in owner user, missing_scope error will occur.
-    #
-    # You can test your token via curl:
-    #   curl -X POST 'https://YOUR-SLACK-TEAM.slack.com/api/users.admin.invite' \
-    #   --data 'email=EMAIL&token=TOKEN&set_active=true' \
-    #   --compressed
-    SLACK_TOKEN: <token>
-
-    # an optional security measure - if it is set, then that token will be required to get invited.
-    # INVITE_TOKEN: <token>
-    
-    LOCALE: en
