+++ package.json
@@ -3,20 +3,14 @@
   "version": "0.3.4",
   "private": true,
   "scripts": {
-    "build": "npm install",
-    "start": "node ./bin/www"
+    "build": "echo 'hello world'"
   },
   "dependencies": {
     "body-parser": "^1.18.0",
     "cookie-parser": "^1.4.0",
-    "debug": "^3.1.0",
-    "dotenv": "^5.0.1",
     "express": "^4.13.0",
-    "i18n": "^0.8.3",
-    "morgan": "^1.6.0",
     "pug": "^2.0.0-rc.4",
     "request": "^2.62.0",
-    "serve-favicon": "^2.3.0",
     "sanitize": "^2.1.0"
   }
 }
